2.1

+++ src/kg2019examples_task4threedimensions/math/Vector2.java
@@ -1,19 +1,19 @@
 package kg2019examples_task4threedimensions.math;
 
 public class Vector2 {
-    private final double p1;
-    private final double p2;
+    private final float p1;
+    private final float p2;
 
-    public Vector2(double p1, double p2) {
+    public Vector2(float p1, float p2) {
         this.p1 = p1;
         this.p2 = p2;
     }
 
-    public double getP1() {
+    public float getP1() {
         return p1;
     }
 
-    public double getP2() {
+    public float getP2() {
         return p2;
     }
 }
